+++ src/components/onboarding/SignUpForm.js
@@ -1,0 +1,71 @@
+import React from 'react';
+import {Field, reduxForm, focus} from 'redux-form';
+import {registerUser} from '../../actions/users';
+import {login} from '../../actions/auth';
+import Input from '../common/input';
+import {required, nonEmpty, matches, length, isTrimmed} from '../common/validators';
+
+
+
+const passwordLength = length({min: 8, max: 72});
+const matchesPassword = matches('password');
+
+export class RegistrationForm extends React.Component {
+    onSubmit(values) {
+
+        const {username, password, firstName, lastName} = values;
+        const user = {username, password, firstName, lastName};
+        return this.props
+            .dispatch(registerUser(user))
+            .then(() => this.props.dispatch(login(username, password)));
+    }
+
+    render() {
+
+        return (
+            <form
+                className="registration-form"
+                onSubmit={this.props.handleSubmit(values =>
+                    this.onSubmit(values)
+                )}>
+                <label className = "registration-form-label" htmlFor="firstName">First name:</label>
+                <Field component={Input} type="text" name="firstName" />
+                <label className = "registration-form-label" htmlFor="lastName">Last name:</label>
+                <Field component={Input} type="text" name="lastName" />
+                <label className = "registration-form-label" htmlFor="username">Username:</label>
+                <Field
+                    component={Input}
+                    type="text"
+                    name="username"
+                    validate={[required, nonEmpty, isTrimmed]}
+                />
+                <label className = "registration-form-label" htmlFor="password">Password:</label>
+                <Field
+                    component={Input}
+                    type="password"
+                    name="password"
+                    validate={[required, passwordLength, isTrimmed]}
+                />
+                <label className = "registration-form-label" htmlFor="passwordConfirm">Confirm password:</label>
+                <Field
+                    component={Input}
+                    type="password"
+                    name="passwordConfirm"
+                    validate={[required, nonEmpty, matchesPassword]}
+                />
+                <button
+                    className = "registration-form-button"
+                    type="submit"
+                    disabled={this.props.pristine || this.props.submitting}>
+                    Register
+                </button>
+            </form>
+        );
+    }
+}
+
+export default reduxForm({
+    form: 'registration',
+    onSubmitFail: (errors, dispatch) =>
+        dispatch(focus('registration', Object.keys(errors)[0]))
+})(RegistrationForm);
